--- drivers/stm32_eth/Makefile
@@ -1,1 +1,0 @@
-include $(RIOTBASE)/Makefile.base
