--- drivers/stm32_eth/doc.txt
@@ -1,13 +1,0 @@
-/*
- *
- * This file is subject to the terms and conditions of the GNU Lesser
- * General Public License v2.1. See the file LICENSE in the top level
- * directory for more details.
- */
-
-/**
-@defgroup    drivers_stm32_common Driver for stm32 ethernet
-@ingroup     drivers_netdev
-@brief       Device Driver for STM32 Ethernet
-
-*/
